requirements.txt
playwright>=1.40
python-docx>=1.1.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0
